gadget,overlord: move OnDiskVolumeFromGadgetVol to a better place
