+++ gadget/install/install.go
@@ -460,65 +460,6 @@ func laidOutStructureForDiskStructure(laidVols map[string]*gadget.LaidOutVolume,
 	return nil, fmt.Errorf("cannot find laid out structure for %q", onDiskStruct.Name)
 }
 
-// OnDiskVolumeFromGadgetVol returns the disk volume matching a gadget volume
-// that has the Device field set, which implies that this should be called only
-// in the context of an installer that set the device in the gadget and
-// returned it to snapd.
-func OnDiskVolumeFromGadgetVol(vol *gadget.Volume) (*gadget.OnDiskVolume, error) {
-	var diskVol *gadget.OnDiskVolume
-	for _, vs := range vol.Structure {
-		if vs.Device == "" || vs.Role == "mbr" || vs.Type == "bare" {
-			continue
-		}
-
-		partSysfsPath, err := sysfsPathForBlockDevice(vs.Device)
-		if err != nil {
-			return nil, err
-		}
-
-		// Volume needs to be resolved only once
-		diskVol, err = onDiskVolumeFromPartitionSysfsPath(partSysfsPath)
-		if err != nil {
-			return nil, err
-		}
-		break
-	}
-
-	if diskVol == nil {
-		return nil, fmt.Errorf("volume %q has no device assigned", vol.Name)
-	}
-
-	return diskVol, nil
-}
-
-// sysfsPathForBlockDevice returns the sysfs path for a block device.
-var sysfsPathForBlockDevice = func(device string) (string, error) {
-	syfsLink := filepath.Join("/sys/class/block", filepath.Base(device))
-	partPath, err := os.Readlink(syfsLink)
-	if err != nil {
-		return "", fmt.Errorf("cannot read link %q: %v", syfsLink, err)
-	}
-	// Remove initial ../../ from partPath, and make path absolute
-	return filepath.Join("/sys/class/block", partPath), nil
-}
-
-// onDiskVolumeFromPartitionSysfsPath creates an OnDiskVolume that
-// matches the disk that contains the given partition sysfs path
-func onDiskVolumeFromPartitionSysfsPath(partPath string) (*gadget.OnDiskVolume, error) {
-	// Removing the last component will give us the disk path
-	diskPath := filepath.Dir(partPath)
-	disk, err := disks.DiskFromDevicePath(diskPath)
-	if err != nil {
-		return nil, fmt.Errorf("cannot retrieve disk information for %q: %v", partPath, err)
-	}
-	onDiskVol, err := gadget.OnDiskVolumeFromDisk(disk)
-	if err != nil {
-		return nil, fmt.Errorf("cannot retrieve on disk volume for %q: %v", partPath, err)
-	}
-
-	return onDiskVol, nil
-}
-
 // applyOnDiskStructureToLaidOut finds the on disk structure from a
 // partition node and takes the laid out information from laidOutVols
 // and inserts it there.
@@ -566,7 +507,7 @@ func WriteContent(onVolumes map[string]*gadget.Volume, allLaidOutVols map[string
 
 	var onDiskVols []*gadget.OnDiskVolume
 	for volName, vol := range onVolumes {
-		onDiskVol, err := OnDiskVolumeFromGadgetVol(vol)
+		onDiskVol, err := gadget.OnDiskVolumeFromGadgetVol(vol)
 		if err != nil {
 			return nil, err
 		}
@@ -694,7 +635,7 @@ func EncryptPartitions(onVolumes map[string]*gadget.Volume, encryptionType secbo
 		parts: make(map[string]partEncryptionData),
 	}
 	for volName, vol := range onVolumes {
-		onDiskVol, err := OnDiskVolumeFromGadgetVol(vol)
+		onDiskVol, err := gadget.OnDiskVolumeFromGadgetVol(vol)
 		if err != nil {
 			return nil, err
 		}
