+++ gadget/ondisk.go
@@ -21,6 +21,8 @@ package gadget
 
 import (
 	"fmt"
+	"os"
+	"path/filepath"
 
 	"github.com/snapcore/snapd/gadget/quantity"
 	"github.com/snapcore/snapd/logger"
@@ -187,4 +189,71 @@ func OnDiskStructureFromPartition(p disks.Partition) (OnDiskStructure, error) {
 		Size:             quantity.Size(p.SizeInBytes),
 		Node:             p.KernelDeviceNode,
 	}, nil
+}
+
+// OnDiskVolumeFromGadgetVol returns the disk volume matching a gadget volume
+// that has the Device field set, which implies that this should be called only
+// in the context of an installer that set the device in the gadget and
+// returned it to snapd.
+func OnDiskVolumeFromGadgetVol(vol *Volume) (*OnDiskVolume, error) {
+	var diskVol *OnDiskVolume
+	for _, vs := range vol.Structure {
+		if vs.Device == "" || vs.Role == "mbr" || vs.Type == "bare" {
+			continue
+		}
+
+		partSysfsPath, err := sysfsPathForBlockDevice(vs.Device)
+		if err != nil {
+			return nil, err
+		}
+
+		// Volume needs to be resolved only once
+		diskVol, err = onDiskVolumeFromPartitionSysfsPath(partSysfsPath)
+		if err != nil {
+			return nil, err
+		}
+		break
+	}
+
+	if diskVol == nil {
+		return nil, fmt.Errorf("volume %q has no device assigned", vol.Name)
+	}
+
+	return diskVol, nil
+}
+
+// sysfsPathForBlockDevice returns the sysfs path for a block device.
+var sysfsPathForBlockDevice = func(device string) (string, error) {
+	syfsLink := filepath.Join("/sys/class/block", filepath.Base(device))
+	partPath, err := os.Readlink(syfsLink)
+	if err != nil {
+		return "", fmt.Errorf("cannot read link %q: %v", syfsLink, err)
+	}
+	// Remove initial ../../ from partPath, and make path absolute
+	return filepath.Join("/sys/class/block", partPath), nil
+}
+
+// onDiskVolumeFromPartitionSysfsPath creates an OnDiskVolume that
+// matches the disk that contains the given partition sysfs path
+func onDiskVolumeFromPartitionSysfsPath(partPath string) (*OnDiskVolume, error) {
+	// Removing the last component will give us the disk path
+	diskPath := filepath.Dir(partPath)
+	disk, err := disks.DiskFromDevicePath(diskPath)
+	if err != nil {
+		return nil, fmt.Errorf("cannot retrieve disk information for %q: %v", partPath, err)
+	}
+	onDiskVol, err := OnDiskVolumeFromDisk(disk)
+	if err != nil {
+		return nil, fmt.Errorf("cannot retrieve on disk volume for %q: %v", partPath, err)
+	}
+
+	return onDiskVol, nil
+}
+
+func MockSysfsPathForBlockDevice(f func(device string) (string, error)) (restore func()) {
+	old := sysfsPathForBlockDevice
+	sysfsPathForBlockDevice = f
+	return func() {
+		sysfsPathForBlockDevice = old
+	}
 }
